Move MockRouter to router module
